% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/have_mat.R
\name{Avb_fun}
\alias{Avb_fun}
\title{Function to check whether Matlab is available}
\usage{
Avb_fun()
}
\value{
a boolean variable
}
\description{
Function to check whether Matlab is available
}
\examples{
Avb_fun()

}
