% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/have_mat.R
\name{avlb_fun}
\alias{avlb_fun}
\title{Function to check whether Matlab is available}
\usage{
avlb_fun()
}
\value{
a boolean variable
}
\description{
Function to check whether Matlab is available
}
\examples{
avlb_fun()

}
