% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mat.R
\name{check_mat}
\alias{check_mat}
\title{Function to check whether the DICOM matrix has been saved successfully}
\usage{
check_mat(file_name)
}
\arguments{
\item{file_name}{the name of the .csv file saved under form_mat_fun()}
}
\value{
a boolean variable
}
\description{
Function to check whether the DICOM matrix has been saved successfully
}
