% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_mat.R
\name{form_mat_fun}
\alias{form_mat_fun}
\title{Function to construct and save data matrix from the DICOM file}
\usage{
form_mat_fun(path_in, out_name)
}
\arguments{
\item{path_in}{afile path for the DICOM variable, needs to be a string}

\item{out_name}{name of the saved .csv file, need to be a string}
}
\value{
saved .csv file of the data matrix
}
\description{
Function to construct and save data matrix from the DICOM file
}
